+++ src/app/locales/en/utils/token-actions/clean-obsolete-tokens-action.js
@@ -2,6 +2,6 @@ export default {
   title: 'Clean up obsolete tokens',
   tipSomethingToClean: 'Clean up obsolete tokens',
   tipNothingToClean: 'Clean up obsolete tokens (nothing to clean)',
-  successNotificationText: 'Selected tokens has been removed.',
+  successNotificationText: 'Selected tokens have been removed.',
   failureNotificationActionName: 'removing tokens',
 };